changing types and adding tracking for wich pin change has been made in case of non unique irqs

+++ kernel_modules/common_helpers/common_helpers.c
@@ -1,7 +1,7 @@
 #include "common_helpers.h"
 #include <linux/kernel.h>
 
-int request_gpios(struct gpio* gpios, const int length) {
+int request_gpios(struct gpio* gpios, const u8 length) {
 	int request_error = gpio_request_array(gpios, length);
 	if(request_error) {
 		printk(KERN_ERR "Unable to request GPIOs!\nCalling gpio_request_array() returned %d\n", request_error);
